+++ assets/js/kirbytags-mode.js
@@ -1,7 +1,7 @@
 /**
  * Visual Markdown Editor Field for Kirby 2
  *
- * @version   1.3.1
+ * @version   1.3.2
  * @author    Jonas Döbertin <hello@jd-powered.net>
  * @copyright Jonas Döbertin <hello@jd-powered.net>
  * @link      https://github.com/JonasDoebertin/kirby-visual-markdown
@@ -51,7 +51,12 @@ CodeMirror.defineMode('kirbytext', function(config, modeConfig) {
         value: [
             // Match a Kirbytext tags attribute value
             {
-                regex: /[^\)]+?(?= (?:[a-z0-9]+:)|\))/i,
+                regex: /[^\)]+? (?=(?:[a-z0-9]+:))/i,
+                token: 'kirbytext-value',
+                next:  'attribute'
+            },
+            {
+                regex: /[^\)]+?(?=\))/i,
                 token: 'kirbytext-value',
                 next:  'attribute'
             },