+++ languages/de.php
@@ -8,6 +8,7 @@ return array(
     'action.name.header2'          => 'h',
     'action.tooltip.bold'          => 'Fett',
     'action.tooltip.italicize'     => 'Kursiv',
+    'action.tooltip.strikethrough' => 'Durchgestrichen',
     'action.tooltip.blockquote'    => 'Zitat',
     'action.tooltip.unorderedList' => 'Unsortierte Liste',
     'action.tooltip.orderedList'   => 'Sortierte Liste',
@@ -17,11 +18,31 @@ return array(
     'action.tooltip.fullscreen'    => 'Vollbildmodus ein-/ausschalten',
 
     /* Help icon subitems */
-    'action.tooltip.markdownLink'  => 'Alles wissenswerte über Markdown',
-    'action.name.markdownLink'     => 'Markdown Hilfe',
-    'action.tooltip.kirbytextLink' => 'Textauszeichnung mit Kirbytext',
-    'action.name.kirbytextLink'    => 'Kirbytext Hilfe',
-    'action.tooltip.issuesLink'    => 'Einen Fehler auf Github melden',
-    'action.name.issuesLink'       => 'Einen Fehler melden',
+    'action.tooltip.shortcutsModal' => 'Übersicht aller Tastenkürzel',
+    'action.name.shortcutsModal'    => 'Tastenkürzel',
+    'action.tooltip.markdownLink'   => 'Alles wissenswerte über Markdown',
+    'action.name.markdownLink'      => 'Markdown Hilfe',
+    'action.tooltip.kirbytextLink'  => 'Textauszeichnung mit Kirbytext',
+    'action.name.kirbytextLink'     => 'Kirbytext Hilfe',
+    'action.tooltip.issuesLink'     => 'Einen Fehler auf Github melden',
+    'action.name.issuesLink'        => 'Einen Fehler melden',
+    'action.tooltip.licenseLink'    => 'Unterstütze die zukünftige Weiterentwicklung',
+    'action.name.licenseLink'       => 'Eine Lizenz kaufen',
+
+
+    /* Shortcuts modal texts */
+    'modal.shortcuts.title'                => 'Tastenkürzel',
+    'modal.shortcuts.heading.action'       => 'Aktion',
+    'modal.shortcuts.heading.shortcut'     => 'Kürzel',
+    'modal.shortcuts.action.heading'       => 'Titel',
+    'modal.shortcuts.action.subheading'    => 'Untertitel',
+    'modal.shortcuts.action.bold'          => 'Fett',
+    'modal.shortcuts.action.italic'        => 'Kursiv',
+    'modal.shortcuts.action.strikethrough' => 'Durchgestrichen',
+    'modal.shortcuts.action.blockquote'    => 'Zitat',
+    'modal.shortcuts.action.unorderedList' => 'Unsortierte Liste',
+    'modal.shortcuts.action.orderedList'   => 'Sortierte Liste',
+    'modal.shortcuts.action.image'         => 'Bild',
+    'modal.shortcuts.action.link'          => 'Link',
 
 );