+++ markdown.php
@@ -2,7 +2,7 @@
 /**
  * Visual Markdown Editor Field for Kirby 2
  *
- * @version   1.3.1
+ * @version   1.3.2
  * @author    Jonas Döbertin <hello@jd-powered.net>
  * @copyright Jonas Döbertin <hello@jd-powered.net>
  * @link      https://github.com/JonasDoebertin/kirby-visual-markdown
@@ -32,13 +32,13 @@ class MarkdownField extends InputField {
     public static $assets = array(
         'js' => array(
             'screenfull-2.0.0.min.js',
-            'codemirror-compressed-5.2.0.min.js',
+            'codemirror-compressed-5.3.0.min.js',
             'kirbytags-mode.js',
             'visualmarkdownfield.js',
             'visualmarkdowneditor.js',
         ),
         'css' => array(
-            'codemirror-5.2.0.css',
+            'codemirror-5.3.0.css',
             'visualmarkdown.css',
         ),
     );
@@ -82,6 +82,7 @@ class MarkdownField extends InputField {
         'header2',
         'bold',
         'italic',
+        'strikethrough',
         'blockquote',
         'unorderedList',
         'orderedList',
@@ -269,6 +270,9 @@ class MarkdownField extends InputField {
      */
     public function input()
     {
+        // Set up modals
+        $modals = tpl::load(__DIR__ . DS . 'partials' . DS . 'modals.php', array('field' => $this));
+
         // Set up translation
         $translation = tpl::load(__DIR__ . DS . 'partials' . DS . 'translation.php', array('translations' => $this->translation));
 
@@ -292,7 +296,10 @@ class MarkdownField extends InputField {
         $wrapper->addClass('markdownfield-wrapper');
         $wrapper->addClass('markdownfield-field-' . $this->name);
 
-        return $wrapper->append($translation)->append($input);
+        return $wrapper
+            ->append($modals)
+            ->append($translation)
+            ->append($input);
     }
 
     /**
@@ -307,6 +314,24 @@ class MarkdownField extends InputField {
         $element = parent::element();
         $element->addClass('field-with-visualmarkdown');
         return $element;
+    }
+
+    /**************************************************************************\
+    *                                 HELPERS                                  *
+    \**************************************************************************/
+
+    /**
+     * Return a translation from the internal translation storage
+     *
+     * @since 1.3.2
+     *
+     * @param  string $key
+     * @param  string $default
+     * @return string
+     */
+    public function lang($key, $default = '')
+    {
+        return (isset($this->translation[$key]) ? $this->translation[$key] : $default);
     }
 
 }